Add ability to "handle" moving goal
This is done by changing the heuristic function, to where it now
changes cost_to_goal depending on how much the goal is expected to
move while it's looking.

It's ugly, but at least it gives a result that works for this
particular task...

NOTE: A* is NOT typically well-equipped to handle a moving goal.
This required a whole deal of bastardization of code, and I'm not
particularily proud of it. I'm also not sure if this would make the
code work for any other task where the goal is moving...